+++ src/AlarmClock/SerialCLI.cpp
@@ -89,6 +89,10 @@ void SerialCLIClass::loop(DateTime __time)
             char *status = strstr(_Serial_buffer, "lamp");
             _print_error(_set_lamp(status));
         }
+        else if (strstr(_Serial_buffer, "inh") != NULL) {
+            char *status = strstr(_Serial_buffer, "inh");
+            _print_error(_set_inh(status));
+        }
         else if (!strcmp(_Serial_buffer, "ls")) {
             _print_error(_list_selected_alarm());
         }
@@ -154,7 +158,9 @@ void SerialCLIClass::loop(DateTime __time)
 SerialCLIClass::SerialCLIClass(AlarmClass *__alarms, void(*__writeEEPROM)(),
                                RTC_DS3231 *__rtc,
                                PWMDimmerClass * __ambientDimmer,
-                               void(*__lamp)(boolean), boolean(*__get_lamp)())
+                               void(*__lamp)(boolean), boolean(*__get_lamp)(),
+                               void(*__set_inhibit)(boolean),
+                               boolean(*__get_inhibit)())
 {
     _alarms = __alarms;
     _writeEEPROM = __writeEEPROM;
@@ -162,6 +168,8 @@ SerialCLIClass::SerialCLIClass(AlarmClass *__alarms, void(*__writeEEPROM)(),
     _ambientDimmer = __ambientDimmer;
     _lamp = __lamp;
     _get_lamp = __get_lamp;
+    _set_inhibit = __set_inhibit;
+    _get_inhibit = __get_inhibit;
 
     strcpy(_prompt, _prompt_default);
 }
@@ -178,6 +186,10 @@ void SerialCLIClass::_print_help()
     Serial.println(F("lamp - get 0|1"));
     _indent(1);
     Serial.println(F("lamp{l} - set 0|1"));
+    _indent(1);
+    Serial.println(F("inh - get inhibit 0|1"));
+    _indent(1);
+    Serial.println(F("inh{i} - set inhibit 0|1"));
     _indent(1);
     Serial.println(F("sel{i} - select alarm{i}"));
     _indent(1);
@@ -284,6 +296,18 @@ SerialCLIClass::error_t SerialCLIClass::_set_lamp(char *status)
         return 0;
     }
     _lamp(_strbyte(status));
+    return 0;
+}
+
+SerialCLIClass::error_t SerialCLIClass::_set_inh(char *status)
+{
+    status = _find_next_digit(status);
+    if (*status == '\0') {
+        Serial.print(F("inhibit: "));
+        Serial.println(_get_inhibit());
+        return 0;
+    }
+    _set_inhibit(_strbyte(status));
     return 0;
 }