+++ src/AlarmClock/SerialCLI.h
@@ -26,6 +26,8 @@ protected:
     PWMDimmerClass * _ambientDimmer;
     void(*_lamp)(boolean);
     boolean(*_get_lamp)();
+    void(*_set_inhibit)(boolean);
+    boolean(*_get_inhibit)();
 
     const char _prompt_default[2 + 1] = "> ";
     char _Serial_buffer[Serial_buffer_length + 1]; // +1 for termination
@@ -50,6 +52,7 @@ protected:
     // commands
     error_t _set_ambient(char *duty);
     error_t _set_lamp(char *status);
+    error_t _set_inh(char *status);
     error_t _select_alarm(byte index);
     error_t _list_selected_alarm();
     error_t _set_enabled(AlarmEnabled __en);
@@ -66,7 +69,8 @@ public:
     void loop(DateTime time);
     SerialCLIClass(AlarmClass *alarms, void(*writeEEPROM)(), RTC_DS3231 *rtc,
                    PWMDimmerClass *ambientDimmer, void(*lamp)(boolean),
-                   boolean(*get_lamp)());
+                   boolean(*get_lamp)(), void(*set_inhibit)(boolean),
+                   boolean(*get_inhibit)());
 };
 
 #endif